update rev1.1231 - hotfixfive
no i am not pilling up hot fixes on the same revision. yes i am. i did a couple of modifications in order to not loop the same obstacles for every nodes during the grid creation so now i just remove the obstacles after they are found and that the node is set unwalkable. Also, i added another fix in the script sccspathfindupdatemain.cs where the loop for retracing the path was also not breaking so i added a switch there so that the condition when a target node is found to break the loop. Now the memory clear doesn't happen every frames... i didn't think it was in my earlier releases. Now, the pathfind arrays are not initialised unless the npc is within range of the pathfind, and if the player is getting out of range, the arrays are clearing.

Also i changed the Stack for a itsalmostastack  //https://stackoverflow.com/questions/748387/how-to-remove-a-stack-item-which-is-not-on-the-top-of-the-stack-in-c-sharp

as the user Mars24 had reported that using a Stack wasn't compiling when using the Cryofall editor.

nine