rev0.1000 - i uploaded the assets that can be multiplied.
rev0.1000 - i uploaded the assets that can be multiplied. Inside of the folder "AssetsToMultiply" are the skeleton.mpk and zombie.mpk and zones.mpk. If you want to create variants of the zombie and skeleton and automatically set zones for those variant mobs, copy the skeleton.mpk and zombie.mpk and zones.mpk inside of the folder #Cryofallscriptgen\original. Then you start the application sccsr16 and the scripts and PNGS and animations will be multiplied with 9 zombie variants and 9 skeleton variants of different color/hp/speed/height. I didn't have the time to set the health bar height yet though. Once the files are multiplied, about 30 secs++ maybe, you can take the .mpks from the multiplied folder and deposit them in your mods folder of Cryofall in your Documents folder. Also, don't forget to get the sccspathfind.mpk that you also need to make this work.