update - rev1.00031
The mobs might have had a different Back/Front color everytime as i was using random values independantly for both sides, so i fixed it so now both sides Back/Front have the same color gradient.
The zombies and zombie variants and all other variants werent set in the script sccspathfindarrays.cs so i added the zombie to the list and also i changed the script a bit so that the variants would be found and set just as their original mob type and template is.
I removed the original mobs spawn in the original zone scripts in order to only use this program's automatic zone mpk creator to setup the zones.
I fixed the first and last zones for mobs spawn as they weren't fixed correctly.
Nine